Improve handling around performance navigation entry (#74)
Older browsers do not have support for PerformanceNavigationTiming. This changes adds fallback for getting navigation type from the deprecated `window.performance.navigation.type` when PerformanceNavigationTiming is not available.

Since `activationStart` is also on PerformanceNavigationTiming, we move this into a helper in the same file as well.

Finally, we can override the built-in browser declarations to get better typing for `getEntriesByType`, `activationStart`, and `document.prerendering`. This allows us to remove the override for `@typescript-eslint/ban-ts-comment` eslint rule since the default is pretty reasonable (`ts-expect-error` is allowed with description) https://typescript-eslint.io/rules/ban-ts-comment/