Soft navigation on a prerendered page should still have `script` navigation type (#76)
This change fixes two issues with the prerender support:
* Soft navigation on a prerendered page should still have `script` navigation type
* Instead of restarting measuring on `prerenderingchange`, we wait for it to init if the document is currently prerendering. This method is described in https://developer.chrome.com/blog/prerender-pages/#impact-on-analytics

Since we are comparing `start` with `activationStart`, restarting measurement doesn't make sense since that would cause `start` to always be slightly after `activationStart` and therefore cause the initial navigation to a prerendered page to be classified as a soft navigation. Waiting for activation and not restarting measurement works better.

This also adds a new test case to the existing prerender tests to perform a soft navigation after landing on a prerendered page.
Due to the linked issue with Playwright, I manually performed the test case after starting the test server with `yarn run express`.